syz-manager: move seed functionality to pkg/manager
